travis: streamline test definitions
No longer distinguish between the tests appart from their distro and
architecture from travis' point of view.

We must distinguish based on architecture, as we do not yet support
cross-architecture builds. We also split by distro as there is no
benefit to running tests for different distros on the same VM, as
they will not be able to share any chaches, so we might as well
parallellize them.

Tests that apply to the same distro/architecture combo are now
always run on the same VM as to utilize any caching we are able
to do.

Now that local_boot and empty_blueprint tests have been merged,
this will not increase the number of tests currently run on a
given VM, so should not affect the total running time of the
tests.

Signed-off-by: Tom Gundersen <teg@jklm.no>