ci/lint: add integration tag
In the current PRs we have an issue with the linter failing on a dead code
in integration tests. The linter isn't failing on main_test.go file
because it somehow ignores the integration tag. However, it fails on
other files which main_test.go depends on and which have the integration
tag.

This commit tells golangci-lint to always use the integration tag while doing
the inspection.