spec: fix building of the tests
Prior this commit building of the RPM package was failing in mock/koji.
The reason for the failure was that we hadn't set correctly the GOPATH and
GO111MODULE environment variables for the `go test -c` command.

This commit takes a whole new approach to the building of tests. There's now
%gobuildtest macro which should be as close as possible to %gobuild macro.
The only change is that it doesn't call `go build`, but `go test -c`.
Hopefully, this approach will make the test building more prone to changes
in go-rpm-macros.

Also, see the comments above the %gobuildtest macro.
fixup