distro: require passing a filename in Manifest()
1b7cb6c11 changed workers so that they don't have access to the distro
that was used to create a manifest anymore. This means that they also
don't know what the filename of the final image in the store is. That
commit worked around this by picking the only file that resulted in that
build. That was a bit of a hack.

Really, image filenames are a user-facing concept. Internally, we're
better of fixing the name to something well-known.

Add a filename parameter to Distro.Manifest() that is independent of the
output type. Set it to "image" from weldr, osbuild-pipeline, and the rcm
API.

Use that name in the worker as well.

Fixes #348