job: get image name from target
1b7cb6c changed workers so that they don't have access to the distro
that was used to create a manifest anymore. I thought this meant that
they also don't know what the filename of the final image in the store
is. That commit worked around this by picking the only file that
resulted in that build. That was a bit of a hack.

Turns out, the `Target` struct already has an ImageName field. Set that
for the implicit local target and use it in the worker.

Fixes #348