docs: add osbuild-composer(7)
Create a new man-page source for `osbuild-composer(7)`. This will be
the overview man-page for composer, as well as a good entrypoint to
learn about the entire osbuild setup.

For now, just include some basic information, as well as copy the bits
from README.md that make sense.

Contributions welcome! This could be a lot more verbose and helpful.