dnf-json: always check for new metadata
After including the fedora updates repository, we were hitting issues
due to expired metadata. The updates repository requires metadata to
be expired at most every six hours, and the default was to do so every
48h.

On further reflection, it was not clear to me what the right value
here is. The values set in the repo files are the maximums, but
users may want to have the ability to access updated content sooner
than that (especially if they are pushing content to a repo and
immediately pulling into osbuild for testing purposes).

For now, we do like lorax and always check for new metadata, rather
than try to come up with some compromise value or (possibly worse)
expose this to the caller to figure out. I imagine this is something
we eventually will want to revisit.

Signed-off-by: Tom Gundersen <teg@jklm.no>