openapi: add spec
This adds two OpenAPI specs. osbuild-composer exposes the API for
generating images and uploading to the cloud, and osbuild-composer-koji
does the same, but specifically for uploading to koji.

The assumption is that the koji images are generic ones, used to be
released, and as such do not require the same level of customization as
the generic API. Moreover, koji will require Kerberos integration, which
the generic API will not.

Signed-off-by: Tom Gundersen <teg@jklm.no>