workflow: enable updates-testing repo when installing osbuild on f31
This is the same fix as in 7ff15e1f, but for the old package (and therefore
for f31).

Copy of the 7ff15e1f commit message:

The source of issue here is that our release cycles are synchronized and
if we want to use the latest released osbuild we need to wait for the
bodhi process. It makes sense that we don't want to use latest osbuild
from git master but we should be confident enough in the version we
submit to bodhi.

Using the version from updates-testing therefore fixes the issue with
synchronization and avoids using unreleased versions.