🧹 Clean up cruft from old jobs
Delete any leftover `/tmp` files from previous test runs and turn
on the `cleanup_composer_directories` option to clean up the
osbuild-composer directories during the time the services are stopped
(when ansible-osbuild is about to deploy the new versions of the
services).

This will hopefully avoid running out of disk space on CI nodes.

Signed-off-by: Major Hayden <major@redhat.com>